♻️ refactor: TextProject, TextDefault 로직 중복 해결
