+++ app/src/components/elements/DashboardContentView/Text/TextDefault.tsx
@@ -1,0 +1,21 @@
+import { H3Text } from '@components/common';
+import { Link } from 'react-router-dom';
+
+type TextDefaultProps = {
+  text: string;
+  link?: string;
+};
+
+export const TextDefault = ({ text, link }: TextDefaultProps) => {
+  return (
+    <>
+      {link !== undefined ? (
+        <Link to={link}>
+          <H3Text>{text}</H3Text>
+        </Link>
+      ) : (
+        <H3Text>{text}</H3Text>
+      )}
+    </>
+  );
+};
