+++ app/src/pages/ProfileGeneralTab/contents/LastRegistered.tsx
@@ -5,8 +5,8 @@ import {
   DashboardContentLoading,
   DashboardContentNotFound,
 } from '@components/elements/DashboardContentView/Error';
-import { TextDefault } from '@components/elements/DashboardContentView/TextDefault';
-import { TextProject } from '@components/elements/DashboardContentView/TextProject';
+import { TextDefault } from '@components/elements/DashboardContentView/Text/TextDefault';
+import { TextProject } from '@components/elements/DashboardContentView/Text/TextProject';
 import { DashboardContent } from '@components/templates/DashboardContent';
 import { useParams } from 'react-router-dom';
 
@@ -36,7 +36,7 @@ export const LastRegistered = () => {
   return (
     <DashboardContent title={title}>
       {lastRegistered != null ? (
-        <TextProject text={lastRegistered} />
+        <TextProject projectName={lastRegistered} />
       ) : (
         <TextDefault text="과제 신청 기록이 없어요 😓" />
       )}
