+++ app/src/pages/ProjectDetailPage/contents/EstimateTime.tsx
@@ -6,7 +6,7 @@ import {
   DashboardContentNotFound,
   NoneDash,
 } from '@components/elements/DashboardContentView/Error';
-import { TextDefault } from '@components/elements/DashboardContentView/TextDefault';
+import { TextDefault } from '@components/elements/DashboardContentView/Text/TextDefault';
 import { DashboardContent } from '@components/templates/DashboardContent';
 import { useParams } from 'react-router-dom';
